requirements.txt
python-pptx
googletrans==4.0.0-rc1